requirements.txt
torch>=2.1.0
diffusers>=0.27.0
transformers>=4.38.0,<5.0.0
accelerate>=0.28.0
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pillow>=10.0.0
python-multipart>=0.0.9
